+++ CHANGELOG.md
@@ -137,6 +137,13 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+## 3.0.14
+
+### Resolved Issues
+
+* \#2552 Ensure `InvalidPath` errors are raised when embedded documents try to
+  get paths from a root selector.
+
 ## 3.0.13
 
 ### Resolved Issues